refactor: split `agglayer_types` into multiple modules/files
